Move command classes to Bindings.Commands
