+++ sources/Google.Solutions.IapDesktop.Application/Views/Options/GeneralOptionsViewModel.cs
@@ -23,6 +23,7 @@ using Google.Solutions.IapDesktop.Application.Data;
 using Google.Solutions.IapDesktop.Application.Services.Adapters;
 using Google.Solutions.IapDesktop.Application.Services.Settings;
 using Google.Solutions.Mvvm.Binding;
+using Google.Solutions.Mvvm.Binding.Commands;
 using System;
 using System.Reflection;
