+++ sources/Google.Solutions.IapDesktop.Extensions.Management.Test/Views/EventLog/TestEventLogViewModel.cs
@@ -31,7 +31,7 @@ using Google.Solutions.IapDesktop.Extensions.Management.Data.Logs;
 using Google.Solutions.IapDesktop.Extensions.Management.History;
 using Google.Solutions.IapDesktop.Extensions.Management.Services.Adapters;
 using Google.Solutions.IapDesktop.Extensions.Management.Views.EventLog;
-using Google.Solutions.Mvvm.Commands;
+using Google.Solutions.Mvvm.Binding.Commands;
 using Google.Solutions.Testing.Application.ObjectModel;
 using Google.Solutions.Testing.Application.Test;
 using Moq;
