+++ sources/Google.Solutions.IapDesktop.Extensions.Management/Views/InstanceProperties/InstancePropertiesInspectorViewModel.cs
@@ -28,8 +28,8 @@ using Google.Solutions.IapDesktop.Application.Services.Integration;
 using Google.Solutions.IapDesktop.Application.Services.ProjectModel;
 using Google.Solutions.IapDesktop.Application.Views.Properties;
 using Google.Solutions.IapDesktop.Extensions.Management.Services.Inventory;
+using Google.Solutions.Mvvm.Binding.Commands;
 using Google.Solutions.Mvvm.Cache;
-using Google.Solutions.Mvvm.Commands;
 using System;
 using System.Diagnostics;
 using System.Threading;
