+++ sources/Google.Solutions.IapDesktop.Extensions.Management/Views/SerialOutput/SerialOutputViewModel.cs
@@ -26,8 +26,8 @@ using Google.Solutions.IapDesktop.Application.ObjectModel;
 using Google.Solutions.IapDesktop.Application.Services.Adapters;
 using Google.Solutions.IapDesktop.Application.Services.Integration;
 using Google.Solutions.IapDesktop.Application.Services.ProjectModel;
+using Google.Solutions.Mvvm.Binding.Commands;
 using Google.Solutions.Mvvm.Cache;
-using Google.Solutions.Mvvm.Commands;
 using System;
 using System.Diagnostics;
 using System.Threading;
