+++ sources/Google.Solutions.IapDesktop.Extensions.Shell.Test/Views/SshKeys/TestAuthorizedPublicKeysViewModel.cs
@@ -29,7 +29,7 @@ using Google.Solutions.IapDesktop.Application.Services.ProjectModel;
 using Google.Solutions.IapDesktop.Application.Views.Dialog;
 using Google.Solutions.IapDesktop.Extensions.Shell.Services.Ssh;
 using Google.Solutions.IapDesktop.Extensions.Shell.Views.SshKeys;
-using Google.Solutions.Mvvm.Commands;
+using Google.Solutions.Mvvm.Binding.Commands;
 using Google.Solutions.Testing.Application.ObjectModel;
 using Moq;
 using NUnit.Framework;
