+++ sources/Google.Solutions.IapDesktop.Extensions.Shell/Views/ConnectionSettings/HtmlPageGenerator.cs
@@ -26,7 +26,7 @@ using Google.Solutions.IapDesktop.Application.Services.Adapters;
 using Google.Solutions.IapDesktop.Application.Services.ProjectModel;
 using Google.Solutions.IapDesktop.Application.Views.ProjectExplorer;
 using Google.Solutions.IapDesktop.Extensions.Shell.Services.ConnectionSettings;
-using Google.Solutions.Mvvm.Commands;
+using Google.Solutions.Mvvm.Binding.Commands;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
