+++ sources/Google.Solutions.IapDesktop.Extensions.Shell/Views/SessionViewBase.cs
@@ -23,7 +23,7 @@ using Google.Solutions.IapDesktop.Application.ObjectModel;
 using Google.Solutions.IapDesktop.Application.Services.Integration;
 using Google.Solutions.IapDesktop.Application.Views;
 using Google.Solutions.Mvvm.Binding;
-using Google.Solutions.Mvvm.Commands;
+using Google.Solutions.Mvvm.Binding.Commands;
 using System;
 
 namespace Google.Solutions.IapDesktop.Extensions.Shell.Views
