+++ sources/Google.Solutions.IapDesktop.Extensions.Shell/Views/SshKeys/AuthorizedPublicKeysViewModel.cs
@@ -28,8 +28,8 @@ using Google.Solutions.IapDesktop.Application.Services.ProjectModel;
 using Google.Solutions.IapDesktop.Application.Views.Dialog;
 using Google.Solutions.IapDesktop.Extensions.Shell.Services.Ssh;
 using Google.Solutions.Mvvm.Binding;
+using Google.Solutions.Mvvm.Binding.Commands;
 using Google.Solutions.Mvvm.Cache;
-using Google.Solutions.Mvvm.Commands;
 using System;
 using System.Diagnostics;
 using System.Linq;
