+++ sources/Google.Solutions.Mvvm/Binding/PropertiesViewModel.cs
@@ -20,6 +20,7 @@
 //
 
 using Google.Solutions.Common.Util;
+using Google.Solutions.Mvvm.Binding.Commands;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
