+++ sources/Google.Solutions.Testing.Application/Views/TestMainForm.cs
@@ -25,7 +25,7 @@ using Google.Solutions.IapDesktop.Application.Services.Adapters;
 using Google.Solutions.IapDesktop.Application.Services.Authorization;
 using Google.Solutions.IapDesktop.Application.Services.Integration;
 using Google.Solutions.IapDesktop.Application.Views;
-using Google.Solutions.Mvvm.Commands;
+using Google.Solutions.Mvvm.Binding.Commands;
 using Google.Solutions.Testing.Common.Integration;
 using System;
 using System.Security.Cryptography.X509Certificates;
